Update tailwind to version 4.0.0
