+++ installer/templates/phx_umbrella/apps/app_name_web/config/config.exs
@@ -29,12 +29,11 @@ config :esbuild,
 
 # Configure tailwind (the version is required)
 config :tailwind,
-  version: "3.4.3",
+  version: "4.0.3",
   <%= @web_app_name %>: [
     args: ~w(
-      --config=tailwind.config.js
-      --input=css/app.css
-      --output=../priv/static/assets/app.css
+      --input=assets/css/app.css
+      --output=priv/static/assets/app.css
     ),
-    cd: Path.expand("../apps/<%= @web_app_name %>/assets", __DIR__)
+    cd: Path.expand("../apps/<%= @web_app_name %>", __DIR__)
   ]<% end %>
