+++ installer/templates/phx_web/components/core_components.ex
@@ -523,7 +523,7 @@ defmodule <%= @web_namespace %>.CoreComponents do
   width, height, and background color classes.
 
   Icons are extracted from the `deps/heroicons` directory and bundled within
-  your compiled app.css by the plugin in your `assets/tailwind.config.js`.
+  your compiled app.css by the plugin in `assets/css/tailwind_heroicons.js`.
 
   ## Examples
