+++ integration_test/config/config.exs
@@ -4,6 +4,6 @@ config :phoenix, :json_library, Jason
 
 config :swoosh, api_client: false
 
-config :tailwind, :version, "3.4.3"
+config :tailwind, :version, "4.0.3"
 
 config :phoenix_live_view, enable_expensive_runtime_checks: true
