+++ integration_test/mix.exs
@@ -55,7 +55,7 @@ defmodule Phoenix.Integration.MixProject do
       {:bcrypt_elixir, "~> 3.0"},
       {:argon2_elixir, "~> 4.0"},
       {:pbkdf2_elixir, "~> 2.0"},
-      {:tailwind, "~> 0.2"},
+      {:tailwind, "~> 0.3"},
       {:heroicons,
        github: "tailwindlabs/heroicons",
        tag: "v2.1.1",
